refactor(autocompact): move context_provider out of gptme/providers/
gptme/providers/ clashed with the existing LLM provider naming in docs
and entry-point groups (gptme.providers). Move the context compression
provider interface to gptme/tools/autocompact/context_provider.py where
it logically belongs.

- Add gptme/tools/autocompact/context_provider.py (was gptme/providers/context.py)
- Remove gptme/providers/ package
- Update handlers.py and hook.py to import from .context_provider
- Update tests/test_context_providers.py to use new module path
- Update docs/context.rst API reference to new module path
