--- gptme/providers/__init__.py
@@ -1,1 +1,0 @@
-"""Context and compression provider interfaces for gptme."""
