+++ gptme/tools/autocompact/hook.py
@@ -14,9 +14,9 @@ from typing import TYPE_CHECKING
 from ...hooks import HookType, StopPropagation, trigger_hook
 from ...llm.models import get_default_model
 from ...message import Message, len_tokens
-from ...providers.context import CompressionConfig, get_context_provider
 from ..base import ToolSpec
 from .config import _get_keep_head
+from .context_provider import CompressionConfig, get_context_provider
 from .decision import should_auto_compact
 from .handlers import cmd_compact_handler
 from .resume import _resume_via_llm
